progress[major] - added zone handling part and weapon upgrade handling part
